pyproject.toml
[tool.poetry]
name = "demo-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
SQLAlchemy = "^1.4.28"
aiomysql = "^0.0.22"
mysqlclient = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.16.0"
aiosqlite = "^0.17.0"
httpx = "^0.21.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
